attempt to subtract with overflow
Running the example panics at line 61 because it somehow gets a negative number with the usizes. This just does a checked_sub and sees if it is `some`.

```
thread 'main' panicked at 'attempt to subtract with overflow',
```

I also added a check to see if the location is not equal to the current center location of neighbors before adding it to the result.